pyproject.toml
[project]
name = "transfermarkt-scraping-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Noam Malka",email = "noamalka2903@gmail.com"}
]
package-mode = false
requires-python = ">=3.13"
dependencies = [
    "attrs (==25.3.0)",
    "automat (==25.4.16)",
    "certifi (==2025.8.3)",
    "cffi (==1.17.1)",
    "charset-normalizer (==3.4.3)",
    "constantly (==23.10.4)",
    "cryptography (==45.0.6)",
    "cssselect (==1.3.0)",
    "defusedxml (==0.7.1)",
    "dnspython (==2.7.0)",
    "filelock (==3.18.0)",
    "hyperlink (==21.0.0)",
    "idna (==3.10)",
    "incremental (==24.7.2)",
    "itemadapter (==0.12.1)",
    "itemloaders (==1.3.2)",
    "jmespath (==1.0.1)",
    "lxml (==6.0.0)",
    "packaging (==25.0)",
    "parsel (==1.10.0)",
    "protego (==0.5.0)",
    "pyasn1 (==0.6.1)",
    "pyasn1-modules (==0.4.2)",
    "pycparser (==2.22)",
    "pydispatcher (==2.0.7)",
    "pymongo (==4.14.0)",
    "pyopenssl (==25.1.0)",
    "queuelib (==1.8.0)",
    "requests (==2.32.4)",
    "requests-file (==2.1.0)",
    "scrapy (==2.13.3)",
    "service-identity (==24.2.0)",
    "setuptools (==80.9.0)",
    "tldextract (==5.3.0)",
    "twisted (==25.5.0)",
    "typing-extensions (==4.14.1)",
    "urllib3 (==2.5.0)",
    "w3lib (==2.3.1)",
    "zope-interface (==7.2)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "crochet (>=2.1.1,<3.0.0)"
]


[build-system]
porequires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
